feat(vue): async value components
